Move onnxruntime tests to standard location, refactor tests to not test so many models, and skip models for now that are currently not working.
